feat: replace old PBMLocationManager with the new one
