+++ PrebidMobile/BuildFiles/PrebidMobile.modulemap
@@ -147,7 +147,6 @@ framework module PrebidMobile {
     
     header "NSDictionary+PBMExtensions.h"
     header "NSMutableDictionary+PBMExtensions.h"
-    header "PBMLocationManager.h"
     
     header "PBMInterstitialAdLoader.h"
     header "PBMInterstitialAdLoaderDelegate.h"
