+++ PrebidMobile/BuildFiles/PrebidMobileSwiftHeaders.h
@@ -27,7 +27,6 @@
 #import "PBMORTBRewardedConfiguration.h"
 #import "PBMORTBRewardedReward.h"
 
-#import "PBMLocationManager.h"
 #import "Log+Extensions.h"
 
 #import "PBMInterstitialAdLoaderDelegate.h"
