+++ PrebidMobile/PrebidMobileRendering/AdTypes/AdView/PBMWebView.m
@@ -21,7 +21,6 @@
 #import "PBMAbstractCreative.h"
 #import "PBMFunctions+Private.h"
 #import "PBMInterstitialDisplayProperties.h"
-#import "PBMLocationManager.h"
 #import "PBMMRAIDController.h"
 #import "PBMMRAIDJavascriptCommands.h"
 #import "PBMMacros.h"
