+++ PrebidMobile/PrebidMobileRendering/Networking/Parameters/PBMGeoLocationParameterBuilder.m
@@ -15,7 +15,6 @@
 
 #import "PBMGeoLocationParameterBuilder.h"
 #import "PBMORTB.h"
-#import "PBMLocationManager.h"
 #import "PBMConstants.h"
 #import "PBMMacros.h"
