+++ PrebidMobile/PrebidMobileRendering/Networking/Parameters/PBMParameterBuilderService.h
@@ -32,6 +32,7 @@
 @class PBMLocationManager;
 @class Prebid;
 @class PBMArbitraryORTBParameterBuilder;
+@class PBMLocationManager;
 
 @interface PBMParameterBuilderService : NSObject
